Add internal/LowPriorityDecoratorComposer
