+++ packages/api/src/decorator/index.ts
@@ -1,7 +1,7 @@
 export { default as ActivityDecorator } from './private/ActivityDecorator';
 export { type ActivityDecoratorRequest } from './private/ActivityDecoratorRequest';
 export { default as DecoratorComposer } from './private/DecoratorComposer';
-export { type DecoratorMiddleware, type DecoratorMiddlewareInit, type DecoratorMiddlewareTypes } from './private/types';
+export { type DecoratorMiddleware, type DecoratorMiddlewareInit, type DecoratorMiddlewareTypes } from './types';
 
 // ActivityGroupingDecorator
