+++ packages/api/src/internal.ts
@@ -1,3 +1,4 @@
+import LowPriorityDecoratorComposer from './decorator/internal/LowPriorityDecoratorComposer';
 import useSetDictateState from './hooks/internal/useSetDictateState';
 
-export { useSetDictateState };
+export { LowPriorityDecoratorComposer, useSetDictateState };
