+++ packages/component/src/decorator/index.ts
@@ -1,2 +1,1 @@
-export { default as BuiltInDecorator } from './private/BuiltInDecorator';
 export { default as WebChatDecorator } from './private/WebChatDecorator';
