+++ packages/component/src/decorator/private/WebChatDecorator.tsx
@@ -1,7 +1,6 @@
 import { DecoratorComposer, type DecoratorMiddleware } from 'botframework-webchat-api/decorator';
 import React, { memo, type ReactNode } from 'react';
 
-import createDefaultActivityGroupingDecoratorMiddleware from '../../Middleware/ActivityGrouping/createDefaultActivityGroupingDecoratorMiddleware';
 import BorderFlair from './BorderFlair';
 import BorderLoader from './BorderLoader';
 import WebChatTheme from './WebChatTheme';
@@ -12,8 +11,7 @@ const middleware: readonly DecoratorMiddleware[] = Object.freeze([
     (next => request => (request.livestreamingState === 'completing' ? BorderFlair : next(request))),
   init =>
     init === 'activity border' &&
-    (next => request => (request.livestreamingState === 'preparing' ? BorderLoader : next(request))),
-  ...createDefaultActivityGroupingDecoratorMiddleware()
+    (next => request => (request.livestreamingState === 'preparing' ? BorderLoader : next(request)))
 ]);
 
 type WebChatDecoratorProps = Readonly<{
@@ -23,9 +21,7 @@ type WebChatDecoratorProps = Readonly<{
 function WebChatDecorator({ children }: WebChatDecoratorProps) {
   return (
     <WebChatTheme>
-      <DecoratorComposer middleware={middleware} priority="low">
-        {children}
-      </DecoratorComposer>
+      <DecoratorComposer middleware={middleware}>{children}</DecoratorComposer>
     </WebChatTheme>
   );
 }
